requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0,<1.0.0
faiss-cpu>=1.8.0
pydantic>=2.7.0
python-dotenv>=1.0.1
requests>=2.32.0
sentence-transformers>=3.0.0
streamlit>=1.44.0
fpdf2>=2.8.0
gTTS>=2.5.0
google-generativeai>=0.8.0
